Fixed unexpected font error on startup ( Just ignoring... ). Fixed unexpected sound stop (retry)

+++ src/emu/device.cpp
@@ -434,7 +434,7 @@ attotime device_t::clocks_to_attotime(u64 numclocks) const noexcept
 		return attotime::never;
 	else if (numclocks < m_clock)
 		return attotime(0, numclocks * m_attoseconds_per_clock);
-	else if (numclocks == LONG_MIN && m_clock == -1)
+	else if ((s64)numclocks == LLONG_MIN && (s32)m_clock == -1)
 		return attotime(0, numclocks * m_attoseconds_per_clock);
 	else
 	{
